Convert these functions to accept an odbcConnect object instead of a string.
I think ultimately this will be more flexible, and it lets the entire session
have a single connection, instead of frenetic opening and closing of sessions.

This isn't any less automatic, and it works with my existing connect_to_tbdbplus
function. Don't know if that's a good thing in the long term.