Revise tx plan query to ensure pt's most recent plans are always included.
A patient was turning up as ongoing with a "New Plan in Place"; on review,
she had completed the new plan already. However, because the new plan
happened after the stop date, it was excluded and she was still counted as
ongoing.

Getting the latest plan per person AND THEN subsetting to those that started
before the stop date fixes this. I'm not sure of the full consequences, but
I think this is okay because I calculate treatment initiations separately.
She *does* appear under treatment initiation, but *does not yet* appear under
any treatment completion plot. I think that's okay.