Give the target an N to suppress a ggplot warning.
ggplot was warning that nine records - those comprising the national target
data - were being removed due to missing values (it dropped the points, which
are sized by N, but not the line - which was basically okay, too). Giving it
N = 1 lets it compute a point size (maybe zero would be better) without
changing the plot much.